tests: simplify serde check
